Extract sunxi boot transition emissions
