+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -65,6 +65,7 @@ mod logger;
 mod model;
 mod phases;
 mod state;
+mod sunxi;
 
 use boot_media::{anchor_bytes_stmt, match_boot_media};
 use caps_tables::{acpi_load_body, acpi_prepare_body, memory_detect_body, smbios_prepare_body};
@@ -74,6 +75,7 @@ use logger::install_logger_body;
 use model::{device_provides, BoardCtx};
 use phases::{phase_init_body, PhaseSpec};
 use state::{emit_adapter_new, emit_adapter_struct};
+use sunxi::{boot_media_select_body, load_next_stage_body};
 
 // =======================================================================
 // Public entry point
@@ -1077,316 +1079,6 @@ fn return_to_fel_body(platform: Platform, ctx: &BoardCtx<'_>) -> TokenStream {
         // returns, so the `-> !` contract is satisfied without a
         // trailing `halt()`.
         unsafe { fstart_soc_sunxi::return_to_fel_from_stash() }
-    }
-}
-
-/// Emit the body of `Board::boot_media_select`.
-///
-/// Called by the executor for `CapOp::BootMediaAuto` and the initial
-/// step of `CapOp::LoadNextStage`.  Inspects the hardware boot-source
-/// register (currently Allwinner-sunxi-only), picks whichever
-/// candidate's `media_ids` contains that byte, and records the
-/// selection on `self._boot_media` so subsequent trampolines read
-/// from the right device.  Returns the chosen candidate's
-/// [`DeviceId`](fstart_types::DeviceId) so the executor can fold it
-/// into its `inited` mask after running `init_device`.
-///
-/// # When the real body is emitted
-///
-/// The stage must:
-///
-/// 1. Be sunxi-format (`SocImageFormat::AllwinnerEgon`).
-///    Non-eGON boards do not implement the `boot_media` register and
-///    referencing `fstart_soc_sunxi` would fail to compile if the
-///    crate is not linked into the stage (generic armv7 and riscv /
-///    aarch64 boards).
-/// 2. Declare `Capability::LoadNextStage` or `Capability::BootMedia`
-///    with an AutoDevice medium.  These are the only paths that lead
-///    the executor to `boot_media_select`; stages without them get a
-///    dead-code `todo!()`.
-///
-/// Both conditions together also guarantee that `fstart_soc_sunxi`
-/// is in scope for this stage (sunxi boards enable the feature flag
-/// that pulls it in).
-///
-/// # The emitted body
-///
-/// - Reads the hardware byte via
-///   `fstart_soc_sunxi::boot_media_at(self._egon_sram_base as usize)`.
-/// - Iterates `candidates` linearly and picks the first whose
-///   `media_ids` contains the byte.
-/// - Writes
-///   `self._boot_media = BootMediaState::Block { device_id, offset, size }`
-///   with the candidate's values (today all LoadNextStage candidates
-///   are block devices; if a future capability adds memory-mapped
-///   candidates we'll need an explicit media-kind tag on
-///   `BootMediaCandidate`).
-/// - Returns `Some(candidate.device)` on match, `None` with a
-///   diagnostic log on miss.
-fn boot_media_select_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let uses_boot_media_select = ctx.stage.capabilities.iter().any(|c| {
-        matches!(
-            c,
-            Capability::LoadNextStage { .. } | Capability::BootMedia(_)
-        )
-    });
-    let is_egon = ctx.config.soc_image_format == fstart_types::SocImageFormat::AllwinnerEgon;
-
-    if !uses_boot_media_select || !is_egon {
-        return quote! {
-            // Dead code: this stage does not declare a capability
-            // that reaches `boot_media_select`, or the board is not
-            // a sunxi eGON board.  Emitting the real body would
-            // reference `fstart_soc_sunxi`, which is only linked
-            // into stages that enable the sunxi feature.
-            let _ = candidates;
-            todo!("board_gen::boot_media_select: stage does not use LoadNextStage/BootMediaAuto, \
-                   or board is not sunxi-eGON")
-        };
-    }
-
-    quote! {
-        // SAFETY: reading the BROM-populated eGON header at
-        // `self._egon_sram_base + 0x28`.  The const-initialised
-        // `_egon_sram_base` is the first stage's `load_addr` from RON,
-        // which is where the BROM dropped us and thus where the
-        // header lives.
-        // SAFETY: _egon_sram_base is the BROM entry point where the
-        // eGON header is mapped in SRAM.
-        let _bm = unsafe {
-            fstart_soc_sunxi::boot_media_at(self._egon_sram_base as usize)
-        };
-        fstart_log::info!("boot media detect: {:#x}", _bm);
-        for candidate in candidates {
-            if candidate.media_ids.iter().any(|&id| id == _bm) {
-                self._boot_media = fstart_stage_runtime::BootMediaState::Block {
-                    device_id: candidate.device,
-                    offset: candidate.offset,
-                    size: candidate.size,
-                };
-                return Some(candidate.device);
-            }
-        }
-        fstart_log::error!(
-            "boot_media_select: no candidate matched boot media {:#x}",
-            _bm,
-        );
-        None
-    }
-}
-
-/// Emit the body of `Board::load_next_stage`.
-///
-/// Reads the eGON header to get the next-stage offset and size, reads
-/// the named stage from the current `self._boot_media` device into
-/// its pre-declared load address, serialises a handoff descriptor,
-/// and jumps with handoff.  This is the capstone of the sunxi
-/// bootblock → main stage transition.
-///
-/// # Emitted shape
-///
-/// 1. `match next_stage { ... }` — one arm per multi-stage entry in
-///    RON other than this stage, each arm binding `load_addr` and
-///    `handoff_addr` locals (and halting on unknown stage names).
-/// 2. Reads `ns_ffs_offset` and `ns_size` from
-///    `fstart_soc_sunxi::next_stage_{offset,size}_at(self._egon_sram_base as usize)`.
-/// 3. Validates both are non-zero; halts otherwise.
-/// 4. `match self._boot_media { Block { device_id, offset, .. } => { match device_id { ... } } ... }`
-///    — per-device dispatch using the block-device field pattern
-///    already used by `sig_verify` / `stage_load`.  Each arm calls
-///    `fstart_capabilities::next_stage::read_stage_to_addr(
-///       self.<field>.as_ref()?, name, next_stage, dev_offset, load_addr, ns_size)`.
-/// 5. Serialises the handoff to `handoff_addr` (= `load_addr - 0x1000`)
-///    via `fstart_capabilities::next_stage::serialize_handoff(dram_size,
-///    handoff_addr)`.  `dram_size` prefers `self._handoff` (carried
-///    over from a previous stage when chained), falls back to the
-///    runtime-detected DRAM size from a `DramInit` driver (when the
-///    stage declares one), and finally to `self._dram_size_static`.
-/// 6. `fstart_platform::jump_to_with_handoff(load_addr, handoff_addr as usize)`.
-///
-/// Dead-codes to `todo!()` for stages that don't declare
-/// `LoadNextStage` or don't use the sunxi eGON format.  Validation
-/// upstream already ensures this combination, so the executor never
-/// reaches the stub at runtime.
-fn load_next_stage_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let uses_load_next_stage = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|c| matches!(c, Capability::LoadNextStage { .. }));
-    let is_egon = ctx.config.soc_image_format == fstart_types::SocImageFormat::AllwinnerEgon;
-
-    if !uses_load_next_stage || !is_egon {
-        return quote! {
-            let _ = next_stage;
-            todo!("board_gen::load_next_stage: stage does not use LoadNextStage, \
-                   or board is not sunxi-eGON")
-        };
-    }
-
-    // Collect (stage_name, load_addr) for every stage defined in the
-    // MultiStage layout, so `match next_stage { ... }` can dispatch
-    // to the right load/handoff addresses.  The current stage is
-    // excluded (can't load ourselves), but with only 2-3 stages in
-    // practice the extra code size is negligible.
-    let stage_arms = match &ctx.config.stages {
-        StageLayout::MultiStage(stages) => stages
-            .iter()
-            .enumerate()
-            .map(|(index, s)| {
-                let name = s.name.as_str();
-                let effective_load_addr =
-                    fstart_types::effective_stage_load_addr(ctx.config, index, s);
-                let load_addr = hex_addr(effective_load_addr);
-                let handoff_addr = hex_addr(effective_load_addr.saturating_sub(0x1000));
-                quote! {
-                    #name => (#load_addr, #handoff_addr),
-                }
-            })
-            .collect::<TokenStream>(),
-        _ => quote! {},
-    };
-
-    // Per-device read-and-dispatch arms.  Matches the block-device
-    // arm pattern used by `sig_verify` / `stage_load`, but here the
-    // body calls `read_stage_to_addr` instead of reading the FFS
-    // anchor.
-    let dev_arms = enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-        .filter(|idx| device_provides(ctx, *idx, Service::BlockDevice))
-        .map(|idx| {
-            let dev = &ctx.devices[idx];
-            let field = format_ident!("{}", dev.name.as_str());
-            let id_lit = proc_macro2::Literal::u8_unsuffixed(idx as u8);
-            let dev_name = dev.name.as_str();
-            quote! {
-                #id_lit => {
-                    fstart_capabilities::next_stage::read_stage_to_addr(
-                        self.#field
-                            .as_ref()
-                            .unwrap_or_else(|| fstart_platform::halt()),
-                        #dev_name,
-                        next_stage,
-                        dev_offset,
-                        load_addr,
-                        ns_size,
-                    )
-                    .unwrap_or_else(|_| {
-                        fstart_log::error!(
-                            "FATAL: failed to read stage from {}",
-                            #dev_name,
-                        );
-                        fstart_platform::halt();
-                    });
-                }
-            }
-        });
-
-    // `dram_size_for_handoff_expr` resolves the DRAM size to pass
-    // into the handoff.  Priority:
-    // 1. `self._handoff.dram_size` (non-first stage carrying a
-    //    previous stage's handoff).
-    // 2. The runtime `detected_size_bytes()` from a DramInit driver
-    //    if the stage declares one — the old codegen's approach.
-    // 3. `self._dram_size_static` as the static fallback.
-    let dram_device = ctx.stage.capabilities.iter().find_map(|cap| match cap {
-        Capability::DramInit { device } => Some(device.as_str()),
-        _ => None,
-    });
-    let dram_size_expr = match dram_device {
-        Some(dev_name) => {
-            let dev = format_ident!("{}", dev_name);
-            quote! {
-                self._handoff
-                    .as_ref()
-                    .filter(|h| h.dram_size > 0)
-                    .map(|h| h.dram_size)
-                    .unwrap_or_else(|| {
-                        self.#dev
-                            .as_ref()
-                            .map(|d| d.detected_size_bytes())
-                            .unwrap_or(self._dram_size_static)
-                    })
-            }
-        }
-        None => quote! {
-            self._handoff
-                .as_ref()
-                .filter(|h| h.dram_size > 0)
-                .map(|h| h.dram_size)
-                .unwrap_or(self._dram_size_static)
-        },
-    };
-
-    quote! {
-        // Stage-specific load and handoff addresses.
-        let (load_addr, handoff_addr): (u64, u64) = match next_stage {
-            #stage_arms
-            other => {
-                fstart_log::error!(
-                    "load_next_stage: unknown next-stage name '{}'",
-                    other,
-                );
-                fstart_platform::halt();
-            }
-        };
-
-        // eGON header read (patched by the FFS assembler at image build).
-        let ns_ffs_offset =
-            // SAFETY: _egon_sram_base points to the eGON header in SRAM.
-            unsafe { fstart_soc_sunxi::next_stage_offset_at(self._egon_sram_base as usize) } as u64;
-        let ns_size =
-            // SAFETY: same as above.
-            unsafe { fstart_soc_sunxi::next_stage_size_at(self._egon_sram_base as usize) } as usize;
-        if ns_ffs_offset == 0 || ns_size == 0 {
-            fstart_log::error!("FATAL: eGON header has zero next_stage_offset/size");
-            fstart_platform::halt();
-        }
-
-        // Per-device dispatch against the current boot medium.
-        match self._boot_media {
-            fstart_stage_runtime::BootMediaState::Block { device_id, offset, .. } => {
-                let dev_offset = offset + ns_ffs_offset;
-                match device_id {
-                    #(#dev_arms)*
-                    _ => {
-                        fstart_log::error!(
-                            "load_next_stage: unknown block device id {}",
-                            device_id,
-                        );
-                        fstart_platform::halt();
-                    }
-                }
-            }
-            fstart_stage_runtime::BootMediaState::Mmio { .. } => {
-                fstart_log::error!(
-                    "load_next_stage: boot medium is a memory-mapped region, \
-                     not a block device",
-                );
-                fstart_platform::halt();
-            }
-            fstart_stage_runtime::BootMediaState::None => {
-                fstart_log::error!(
-                    "load_next_stage: no boot medium selected",
-                );
-                fstart_platform::halt();
-            }
-        }
-
-        // Handoff serialisation + jump.  `dram_size` prefers the
-        // runtime-detected value (from handoff or DramInit driver)
-        // and falls back to the RON static size.
-        let dram_size: u64 = #dram_size_expr;
-        fstart_capabilities::next_stage::serialize_handoff(dram_size, handoff_addr)
-            .unwrap_or_else(|_| {
-                fstart_log::error!("FATAL: handoff serialize failed");
-                fstart_platform::halt();
-            });
-        fstart_log::info!(
-            "jumping to stage '{}' at {:#x}",
-            next_stage,
-            load_addr,
-        );
-        fstart_platform::jump_to_with_handoff(load_addr, handoff_addr as usize)
     }
 }
